+++ src/app.css
@@ -91,8 +91,6 @@ button {
 /*
   A full size invisible container to cover the entire screen for modularity purposes, 
   not accidentally trigger click on other elements e.g. accidentally creating a new task on the calendar when you're just trying to exit
-
-  For some reason 100vh doesn't cover the bottom, so I just make it ridiculously big
 */
 .fullscreen-invisible-modular-popup-layer {
   width: 100vw;
